+++ bootstrap.sh
@@ -116,6 +116,11 @@ function initializeGit {
   rm -rf ".git"
   git init >/dev/null
   git add "*" >/dev/null
+
+  echo "Adding commit hooks..."
+  swift run komondor install
+
+  echo "Creating initial commit."
   git commit -m "Initial commit" >/dev/null
 }