+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * Update Rome usage code in Podfile.
 * Re-work settings config structure.
 * Add file lists for quicker build phases.
+* Add commit hooks for running swiftformat & lint autocorrect.
 
 ### 19 March 2020